用户个人商机修改优化版

+++ anchongnet/app/Business_img.php
@@ -68,10 +68,15 @@ class Business_img extends Model implements AuthenticatableContract,
     */
     public function delimg($id)
     {
-        if($this->where('id', '=', $id)->delete()){
-            return true;
+        $data=$this->where('id', '=', $id)->get()->toArray();
+        if($data){
+            if($this->where('id', '=', $id)->delete()){
+                return true;
+            }else{
+                return false;
+            }
         }else{
-            return false;
+            return true;
         }
     }
 }
